requirements.txt
asgiref==3.11.1
black==26.1.0
click==8.3.1
Django==5.0.14
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.1.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
psycopg2-binary==2.9.11
python-dotenv==1.2.2
pytokens==0.4.1
sqlparse==0.5.5
tomli==2.4.0
typing_extensions==4.15.0
whitenoise==6.12.0
